refactor(system): extract beanfun webview popups into a service
commands/system.rs drops from 1555 to 1044 lines. Per Clean Architecture
(webview/cookie side effects belong in services/), moved the cohesive beanfun
popup group into new services/web_popup_service.rs:

- open_gash_popup (top-up), open_member_popup (member center)
- open_customer_service, open_auth_popup, open_web_popup
- web_token_from_jar (shared bfWebToken cookie-jar lookup)

Command handlers are now thin wrappers delegating to the service (State is
unwrapped to &AppState at the boundary). resize_gash_popup stays as a thin
window-resize command. No command names/signatures changed — frontend and
lib.rs registration untouched. Pure mechanical move.
NOTE: webview/cookie popups only verifiable via a live beanfun session —
smoke-test gash/member/customer-service/report popups before merging.
