+++ src-tauri/src/services/mod.rs
@@ -17,5 +17,6 @@ pub mod recaptcha_window;
 pub mod session_key_fallback;
 pub mod update_service;
 pub mod web_launch;
+pub mod web_popup_service;
 pub mod webview_login;
 pub mod webview_util;
